[flang][OpenMP] Move lowering of ATOMIC to separate file, NFC
Reinstate commits e5559ca4 and 925dbc79. Fix the issues with compilation
hangs by including DenseMapInfo specialization where the corresponding
DenseMap was defined.

Ref: https://github.com/llvm/llvm-project/pull/144960
